requirements.txt
appdirs==1.4.3
atomicwrites==1.3.0
attrs==19.1.0
black==19.3b0
Click==7.0
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
more-itertools==7.0.0
pluggy==0.9.0
psycopg2-binary==2.8.2
py==1.8.0
pytest==4.4.0
python-dotenv==0.10.1
six==1.12.0
SQLAlchemy==1.3.2
toml==0.10.0
Werkzeug==0.15.2